requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
SQLAlchemy==2.0.23
Werkzeug==2.3.7
alembic==1.12.0
greenlet==3.0.1